Make Instances a module of its own
